Fix invitation code input issue (#70)
In the invitation code input box, when you start typing something,
it gets automatically erased.

This turned out to be a known issue with the react-native-dialog-input
library:
https://github.com/joseestrella89/react-native-dialog-input/issues/20

The suggested fix involves a patch to the library,
which is luckily just a single file,
so I included the patched version in the source code tree.